% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_communes.R
\docType{data}
\name{border_communes}
\alias{border_communes}
\title{Norwegian cities}
\format{A rda file}
\source{
to be added
}
\description{
Norwegian cities
}
\references{
to be added
}
\keyword{datasets}

